refactored errors, and generalized ControllerAdvice
