+++ src/main/java/com/example/primo_progetto_spring/tutor/service/TutorService.java
@@ -3,7 +3,7 @@ package com.example.primo_progetto_spring.tutor.service;
 
 import com.example.primo_progetto_spring.classroom.entity.Classroom;
 import com.example.primo_progetto_spring.classroom.repository.ClassroomRepository;
-import com.example.primo_progetto_spring.tutor.service.error.MissingFieldMandatoryException;
+import com.example.primo_progetto_spring.errors.MissingFieldMandatoryException;
 import com.example.primo_progetto_spring.student.repository.StudentRepository;
 import com.example.primo_progetto_spring.tutor.entity.Tutor;
 import com.example.primo_progetto_spring.tutor.repository.TutorRepository;
